Add support for using CloudFlare Argo Tunnels
When using CloudFlare Argo Tunnels to expose the local environment(Instead of ngrok for example), the selfCaller function fails if the CloudFlare setting to redirect all HTTP -> HTTPS traffic is enabled. From what I've seen, this is due to the protocol detection using the protocol that was used locally, rather than at the edge.

For example:

Submit Middleware Test on FCC site(Requires HTTPS):
FCC --> CF URL(https://fcc-backend.owls.icu)
CF -- Tunnel --> to local dev env. Local env sees the protocol as HTTP
Local Server --> Request to http://fcc-backend.owls.icu
- selfCaller() sends a request to the <proto><host> which in this example evaluates to http://fcc-backend.owls.icu
CloudFlare responds with a 301 Redirect to https://fcc-backend.owls.icu. Unfortunately, the http package does not support following redirects as far as I can tell.

This change configures the selfCaller() function to force the port and protocol to be https when the x-forwarded-for and cdn-loop headers are present. As this hasn't caused issue for others from what I can see, I limited the scope to just be CloudFlare specific.